Move the sink into util::vec.
